Padronizacao das mensagens de erros

+++ src/ERP.Api/Controllers/ExemploController.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using ERP.Api.ViewModels.Exemplos;
+using ERP.Business.Interfaces;
 using ERP.Business.Interfaces.Exemplos;
 using ERP.Business.Models;
 using Microsoft.AspNetCore.Mvc;
@@ -8,57 +9,66 @@ namespace ERP.Api.Controllers
 {
     [ApiController]
     [Route("api/exemplo")]
-    public class ExemploController : ControllerBase
+    public class ExemploController : MainController<ExemploController>
     {
         private readonly IExemploService _exemploService;
         private readonly IMapper _mapper;
-        public ExemploController(IExemploService exemploService, IMapper mapper)
+        public ExemploController(IExemploService exemploService, 
+            IMapper mapper,
+            IErrorNotifier erroNotifier) : base(erroNotifier)
         {
             _exemploService = exemploService;
             _mapper = mapper;
         }
        
         [HttpPost]
-        public async Task<Exemplo> Add(ExemploCreateViewModel viewModel)
+        public async Task<ActionResult<ExemploCreateViewModel>> Add(ExemploCreateViewModel viewModel)
         {
+            if (!ModelState.IsValid) return CustomResponse(ModelState);
+
             var exemplo = new Exemplo(viewModel.Nome, viewModel.CpfCnpj);
 
             await _exemploService.Add(exemplo);
-            return exemplo;
+            return CustomResponse(viewModel);
         }
         [HttpGet("GetAll")]
         public async Task<IEnumerable<Exemplo>> GetAll()
         {
-            var users = await _exemploService.GetAll();
+            var exemplos = await _exemploService.GetAll();
 
-            return users;
+            return exemplos;
         }
 
         [HttpGet("{id:Guid}")]
-        public async Task<Exemplo> GetById(Guid id)
+        public async Task<ActionResult<Exemplo>> GetById(Guid id)
         {
             var exemplo = await _exemploService.GetById(id);
+            if (exemplo == null) return NotFound();
 
             return exemplo;
         }
 
         [HttpPut]
-        public async Task<Exemplo> Update(ExemploUpdateViewModel viewModel)
+        public async Task<ActionResult<ExemploUpdateViewModel>> Update(ExemploUpdateViewModel viewModel)
         {
+            if (!ModelState.IsValid) return CustomResponse(ModelState);
+
             var exemplo = await _exemploService.GetById(viewModel.Id);
+            if (exemplo == null) return NotFound();
 
-            _mapper.Map<ExemploUpdateViewModel, Exemplo>(viewModel, exemplo);
-
-            await _exemploService.Update(exemplo);
-            return exemplo;
+            await _exemploService.Update(_mapper.Map<Exemplo>(viewModel));
+            return CustomResponse(viewModel);
         }
 
         [HttpDelete("{id:Guid}")]
-        public async Task<ActionResult> Remove(Guid id)
+        public async Task<ActionResult<ExemploViewModel>> Remove(Guid id)
         {
+            var exemplo = await _exemploService.GetById(id);
+            if (exemplo == null) return NotFound();
+
             await _exemploService.Remove(id);
 
-            return Ok();
+            return CustomResponse();
         }
     }
 }
